# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "server",
    "description": "",
    "license": {
      "name": "AGPL-3.0-only",
      "identifier": "AGPL-3.0-only"
    },
    "version": "0.0.1"
  },
  "paths": {
    "/api/artists": {
      "get": {
        "tags": [
          "artists"
        ],
        "operationId": "list_artists",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Page number, 1-indexed. Defaults to 1.",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "per_page",
            "in": "query",
            "description": "Items per page. Defaults to 20. The server enforces a maximum.",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "q",
            "in": "query",
            "description": "Text search by artist name.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "description": "Field to sort by. Defaults to `name`.",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/ArtistSort"
            }
          },
          {
            "name": "sort_dir",
            "in": "query",
            "description": "Sort direction. Defaults to `asc`.",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/SortDir"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of artists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResponse_ArtistSummary"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "artists"
        ],
        "operationId": "create_artist",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateArtistRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created artist",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtistResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      }
    },
    "/api/artists/{id}": {
      "get": {
        "tags": [
          "artists"
        ],
        "operationId": "get_artist",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Artist ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Artist detail",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtistResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "artists"
        ],
        "operationId": "update_artist",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Artist ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateArtistRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Updated artist",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtistResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      },
      "delete": {
        "tags": [
          "artists"
        ],
        "operationId": "delete_artist",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Artist ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      }
    },
    "/api/artists/{id}/images": {
      "post": {
        "tags": [
          "artists"
        ],
        "operationId": "upload_artist_image",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Artist ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/ImageUpload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Image uploaded and linked",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtistImageInfo"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Artist not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      }
    },
    "/api/artists/{id}/images/{asset_id}": {
      "delete": {
        "tags": [
          "artists"
        ],
        "operationId": "delete_artist_image",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Artist ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "asset_id",
            "in": "path",
            "description": "Asset ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      },
      "patch": {
        "tags": [
          "artists"
        ],
        "operationId": "update_artist_image_kind",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Artist ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "asset_id",
            "in": "path",
            "description": "Asset ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateArtistImageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Kind updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtistImageInfo"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ArtistImageInfo": {
        "allOf": [
          {
            "$ref": "#/components/schemas/AssetInfo"
          },
          {
            "type": "object",
            "required": [
              "kind"
            ],
            "properties": {
              "kind": {
                "type": "string",
                "description": "Serialized [`ArtistImageKind`] value indicating the image's semantic role."
              }
            }
          }
        ],
        "description": "An image asset attached to an artist with its semantic role."
      },
...
      "ArtistLinkInfo": {
        "type": "object",
        "description": "An external link associated with an artist.",
        "required": [
          "id",
          "url",
          "kind"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Unique identifier of the link record."
          },
          "kind": {
            "type": "string",
            "description": "Serialized [`ArtistLinkKind`] value indicating the platform."
          },
          "label": {
            "type": [
              "string",
              "null"
            ],
            "description": "Optional display override label."
          },
          "url": {
            "type": "string",
            "description": "Fully qualified URL."
          }
        }
      },
      "ArtistLinkInput": {
        "type": "object",
        "description": "Input for creating an external link on an artist.",
        "required": [
          "url",
          "kind"
        ],
        "properties": {
          "kind": {
            "$ref": "#/components/schemas/ArtistLinkKind",
            "description": "Platform this link points to."
          },
          "label": {
            "type": [
              "string",
              "null"
            ],
            "description": "Optional display override label."
          },
          "url": {
            "type": "string",
            "description": "Fully qualified URL."
          }
        }
      },
...
      "ArtistResponse": {
        "type": "object",
        "description": "A full artist record returned by detail and mutation endpoints.",
        "required": [
          "id",
          "name",
          "images",
          "links",
          "song_count",
          "performance_count"
        ],
        "properties": {
          "description": {
            "type": [
              "string",
              "null"
            ],
            "description": "Optional freeform bio or description."
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Unique identifier."
          },
          "images": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ArtistImageInfo"
            },
            "description": "Images attached to this artist."
          },
          "links": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ArtistLinkInfo"
            },
            "description": "External links for this artist."
          },
          "name": {
            "type": "string",
            "description": "Display name."
          },
          "performance_count": {
            "type": "integer",
            "format": "int64",
            "description": "Number of performances in which this artist appears as a singer.",
            "minimum": 0
          },
          "song_count": {
            "type": "integer",
            "format": "int64",
            "description": "Number of songs for which this artist is credited as an original artist.",
            "minimum": 0
          }
        }
      },
      "ArtistSort": {
        "type": "string",
        "description": "Field to sort artists by in list endpoints.",
        "enum": [
          "name",
          "song_count"
        ]
      },
...
      "AssetInfo": {
        "type": "object",
        "description": "Common fields describing a media asset, whether internally hosted or externally referenced.\n\nExactly one of `storage_url` or `external_url` is non-null per asset.",
        "required": [
          "asset_id"
        ],
        "properties": {
          "asset_id": {
            "type": "string",
            "format": "uuid",
            "description": "Unique identifier of the asset record."
          },
          "credits": {
            "type": [
              "string",
              "null"
            ],
            "description": "Attribution text for the content creator."
          },
          "external_url": {
            "type": [
              "string",
              "null"
            ],
            "description": "Referenced external URL. Present for externally linked content only."
          },
          "source_url": {
            "type": [
              "string",
              "null"
            ],
            "description": "Link to the original source post or stream."
          },
          "storage_url": {
            "type": [
              "string",
              "null"
            ],
            "description": "CDN or file server URL. Present for internally hosted files only."
          },
          "title": {
            "type": [
              "string",
              "null"
            ],
            "description": "Display name or original filename. May be absent for unnamed uploads."
          }
        }
      },
...
      "CreateArtistRequest": {
        "type": "object",
        "description": "Request body for `POST /api/artists`.",
        "required": [
          "name",
          "links"
        ],
        "properties": {
          "description": {
            "type": [
              "string",
              "null"
            ],
            "description": "Optional freeform bio or description."
          },
          "links": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ArtistLinkInput"
            },
            "description": "External links to create alongside the artist."
          },
          "name": {
            "type": "string",
            "description": "Display name."
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "description": "Body returned for all error responses.",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string",
            "description": "Human readable description of the error."
          }
        }
      },
...
      "ImageUpload": {
        "type": "object",
        "description": "Placeholder schema for image multipart upload bodies.",
        "required": [
          "file",
          "kind"
        ],
        "properties": {
          "credits": {
            "type": [
              "string",
              "null"
            ]
          },
          "file": {
            "type": "string",
            "format": "binary"
          },
          "kind": {
            "type": "string",
            "description": "Semantic role of the image. See [`ArtistImageKind`](api_types::artists::ArtistImageKind)."
          },
          "source_url": {
            "type": [
              "string",
              "null"
            ]
          },
          "title": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "PagedResponse_ArtistSummary": {
        "type": "object",
        "description": "Generic envelope returned by all paginated list endpoints.",
        "required": [
          "items",
          "total",
          "page",
          "per_page"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "description": "A lean artist record returned by the list endpoint.",
              "required": [
                "id",
                "name",
                "images",
                "song_count"
              ],
              "properties": {
                "description": {
                  "type": [
                    "string",
                    "null"
                  ],
                  "description": "Optional freeform bio or description."
                },
                "id": {
                  "type": "string",
                  "format": "uuid",
                  "description": "Unique identifier."
                },
                "images": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ArtistImageInfo"
                  },
                  "description": "Images attached to this artist."
                },
                "name": {
                  "type": "string",
                  "description": "Display name."
                },
                "song_count": {
                  "type": "integer",
                  "format": "int64",
                  "description": "Number of songs for which this artist is credited as an original artist.",
                  "minimum": 0
                }
              }
            },
            "description": "Items on the current page."
          },
          "page": {
            "type": "integer",
            "format": "int32",
            "description": "The page that was returned, 1-indexed.",
            "minimum": 0
          },
          "per_page": {
            "type": "integer",
            "format": "int32",
            "description": "Items per page as applied by the server (may be less than requested if capped).",
            "minimum": 0
          },
          "total": {
            "type": "integer",
            "format": "int64",
            "description": "Total number of matching items across all pages.",
            "minimum": 0
          }
        }
      },
...
      "SortDir": {
        "type": "string",
        "description": "Sort direction for list endpoints.",
        "enum": [
          "asc",
          "desc"
        ]
      },
...
      "UpdateArtistImageRequest": {
        "type": "object",
        "description": "Request body for `PATCH /api/artists/{id}/images/{image_id}`.",
        "required": [
          "kind"
        ],
        "properties": {
          "kind": {
            "type": "string",
            "description": "Serialized [`ArtistImageKind`] value to assign."
          }
        }
      },
      "UpdateArtistRequest": {
        "type": "object",
        "description": "Request body for `PUT /api/artists/{id}`.\n\nImages are managed via the `/images` subresource.",
        "required": [
          "name",
          "links"
        ],
        "properties": {
          "description": {
            "type": [
              "string",
              "null"
            ],
            "description": "Optional freeform bio or description."
          },
          "links": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ArtistLinkInput"
            },
            "description": "Replaces all existing external links."
          },
          "name": {
            "type": "string",
            "description": "Display name."
          }
        }
      },
...
    },
    "securitySchemes": {
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session"
      }
    }
  }
}